backend/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-filter==24.3
django-cors-headers==4.6.0
django-mptt==0.16.0
django-jazzmin==3.0.1
django-redis==5.4.0
django-celery-beat==2.7.0
django-extensions==3.2.3
django-import-export==4.3.4
drf-spectacular==0.28.0
psycopg2-binary==2.9.10
redis==5.2.1
celery==5.4.0
Pillow==11.1.0
python-dotenv==1.0.1
requests==2.32.3
whitenoise==6.8.2
gunicorn==23.0.0
beautifulsoup4==4.12.3
lxml==5.3.0
# Optional: AI integrations for PC configurator evaluation
openai>=1.0.0
anthropic>=0.25.0
